+++ signalfx/resource_signalfx_aws_integration_test.go
