+++ signalfx/resource_signalfx_azure_integration_test.go
