+++ signalfx/resource_signalfx_slack_integration_test.go
