+++ website/signalfx.erb
@@ -13,6 +13,12 @@
         <li<%= sidebar_current("docs-signalfx-resource") %>>
           <a href="#">Resources</a>
           <ul class="nav nav-visible">
+            <li<%= sidebar_current("docs-signalfx-resource-aws-integration") %>>
+              <a href="/docs/providers/signalfx/r/aws_integration.html">signalfx_aws_integration</a>
+            </li>
+            <li<%= sidebar_current("docs-signalfx-resource-azure-integration") %>>
+              <a href="/docs/providers/signalfx/r/azure_integration.html">signalfx_azure_integration</a>
+            </li>
             <li<%= sidebar_current("docs-signalfx-resource-dashboard") %>>
               <a href="/docs/providers/signalfx/r/dashboard.html">signalfx_dashboard</a>
             </li>
@@ -22,29 +28,23 @@
             <li<%= sidebar_current("docs-signalfx-resource-detector") %>>
               <a href="/docs/providers/signalfx/r/detector.html">signalfx_detector</a>
             </li>
-            <li<%= sidebar_current("docs-signalfx-resource-aws-integration") %>>
-              <a href="/docs/providers/signalfx/r/integration_aws.html">signalfx_integration_aws</a>
-            </li>
-            <li<%= sidebar_current("docs-signalfx-resource-azure-integration") %>>
-              <a href="/docs/providers/signalfx/r/integration_azure.html">signalfx_integration_azure</a>
-            </li>
-            <li<%= sidebar_current("docs-signalfx-resource-gcp-integration") %>>
-              <a href="/docs/providers/signalfx/r/integration_gcp.html">signalfx_integration_gcp</a>
-            </li>
-            <li<%= sidebar_current("docs-signalfx-resource-pagerduty-integration") %>>
-              <a href="/docs/providers/signalfx/r/integration_pagerduty.html">signalfx_integration_pagerduty</a>
-            </li>
-            <li<%= sidebar_current("docs-signalfx-resource-slack-integration") %>>
-              <a href="/docs/providers/signalfx/r/integration_slack.html">signalfx_integration_slack</a>
-            </li>
             <li<%= sidebar_current("docs-signalfx-resource-event-feed-chart") %>>
               <a href="/docs/providers/signalfx/r/event_feed_chart.html">signalfx_event_feed_chart</a>
+            </li>
+            <li<%= sidebar_current("docs-signalfx-resource-gcp-integration") %>>
+              <a href="/docs/providers/signalfx/r/gcp_integration.html">signalfx_gcp_integration</a>
             </li>
             <li<%= sidebar_current("docs-signalfx-resource-heatmap-chart") %>>
               <a href="/docs/providers/signalfx/r/heatmap_chart.html">signalfx_heatmap_chart</a>
             </li>
             <li<%= sidebar_current("docs-signalfx-resource-list-chart") %>>
               <a href="/docs/providers/signalfx/r/list_chart.html">signalfx_list_chart</a>
+            </li>
+            <li<%= sidebar_current("docs-signalfx-resource-pagerduty-integration") %>>
+              <a href="/docs/providers/signalfx/r/pagerduty_integration.html">signalfx_pagerduty_integration</a>
+            </li>
+            <li<%= sidebar_current("docs-signalfx-resource-slack-integration") %>>
+              <a href="/docs/providers/signalfx/r/slack_integration.html">signalfx_slack_integration</a>
             </li>
             <li<%= sidebar_current("docs-signalfx-resource-org-token") %>>
               <a href="/docs/providers/signalfx/r/org_token.html">signalfx_org_token</a>
